🧐 Code review comments implemented
